🔨 Refactor: 회원가입 api 수정
