+++ src/components/nav-bar/logout-drawer.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 
 import { Button, Drawer } from "@/components/common";
 import { useToast } from "@/hooks";
-import { deleteCookie } from "@/utils/next-cookies";
+import { deleteCookie } from "@/utils/next-cookie";
 
 interface LogoutDrawerProps {
   isOpen: boolean;
