+++ src/lib/api/auth/sign-in.ts
@@ -1,9 +1,7 @@
-"use server";
-
 import axios from "axios";
-import { cookies } from "next/headers";
 
 import instance from "@/lib/api/axios-instance";
+import { setCookie } from "@/utils/next-cookie";
 
 const signIn = async (email: string, password: string) => {
   try {
@@ -12,8 +10,9 @@ const signIn = async (email: string, password: string) => {
     const { data } = response;
 
     if (response.status === 200) {
-      cookies().set("token", data.accessToken);
-      cookies().set("refreshToken", data.refreshToken);
+      await setCookie("token", data.accessToken);
+      await setCookie("refreshToken", data.refreshToken);
+      await setCookie("userId", data.user.id);
 
       return { success: true };
     }
